closes #7, closes #5 : adds all events for parsing and updates sync workers

+++ modules-data/domain/src/main/kotlin/com/bizilabs/streeek/lib/domain/models/AccountDomain.kt
@@ -4,7 +4,7 @@ import kotlinx.datetime.LocalDateTime
 
 data class AccountDomain(
     val id: Long,
-    val githubId: Int,
+    val githubId: Long,
     val username: String,
     val email: String,
     val bio: String,
