+++ modules-sources/local/src/main/kotlin/com/bizilabs/streeek/lib/local/models/AccountCache.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.json.Json
 @Serializable
 data class AccountCache(
     val id: Long,
-    val githubId: Int,
+    val githubId: Long,
     val username: String,
     val email: String,
     val bio: String,
